Move autoremove code to util and add tests
